+++ components/calendar/src/coptic.rs
@@ -76,7 +76,7 @@ impl CalendarArithmetic for Coptic {
         }
     }
 
-    fn months_for_every_year() -> u8 {
+    fn months_for_every_year(_: i32) -> u8 {
         13
     }